requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pycodestyle==2.12.1
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
toml==0.10.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3